Why I hate Hugh Juliana and Mara - A Story
It was dark Tuesday when I entered the programming room. All was well, I was not sleep deprived, I was well caffeinated, and seemed good.  Ahh how i naive i was. I sat down, opened the laptop and began to look at the code my FRIENDS (so called) had done to contribute to the team. Without thought, with trust in my comrades, made the incorrect assumption that they were capable people. Unfortunately y faith was unfounded, as they CHANGED THE STRUCTURE OF THE CODE AND DIDN'T TELL ANYONE.